requirements.txt
PyQt6>=6.4.0
pytesseract>=0.3.10
Pillow>=9.0.0
requests>=2.25.0
# Add pdf2image or PyMuPDF later for PDF support
# Add reportlab or similar later for PDF output 